feat: add configuration and relocate database provider
