+++ .example.env
@@ -1,0 +1,15 @@
+APP_NAME=Logistic
+APP_ENV=local
+APP_DEBUG=false
+
+DB_CONNECTION=postgres
+DB_HOST=localhost
+DB_PORT=5432
+DB_USERNAME=postgres
+DB_PASSWORD=postgres
+DB_SYNCHRONIZE=true
+DB_DEBUG=true
+
+REDIS_HOST=127.0.0.1
+REDIS_PASSWORD=null
+REDIS_PORT=6379
