+++ src/common/database/database.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
 import { databaseProviders } from "./database.providers";
+import { ConfigurationModule } from "../../config/configuration.module";
 
 @Module({
+    imports: [ConfigurationModule],
     providers: [...databaseProviders],
     exports: [...databaseProviders]
 })
